Revert "fix: Change CoreData model to ensure offline files do not have strict ties to their origin. (#165)" (#169)
This reverts commit 18a5fb93d6f4fa3c5cf52467456a22386af29621.
