+++ AmahiAnywhere/AmahiAnywhere/Presentation/Offline/OfflineFilesTableViewController+UITableViewDelegates.swift
@@ -28,11 +28,9 @@ extension OfflineFilesTableViewController {
         if offlineFile.stateEnum != .downloading {
             cell.progressView.isHidden = true
         } else {
-            if let remoteUrl = offlineFile.remoteFileURL() {
-                let keyExists = DownloadService.shared.activeDownloads[remoteUrl] != nil
-                if !keyExists {
-                    offlineFile.stateEnum = .completedWithError
-                }
+            let keyExists = DownloadService.shared.activeDownloads[offlineFile.remoteFileURL()] != nil
+            if !keyExists {
+                offlineFile.stateEnum = .completedWithError
             }
         }
