--- AmahiAnywhere/OfflineFilesModel.xcdatamodeld/OfflineFilesModel v2.xcdatamodel/contents
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<model type="com.apple.IDECoreDataModeler.DataModel" documentVersion="1.0" lastSavedToolsVersion="14490.98" systemVersion="18E226" minimumToolsVersion="Automatic" sourceLanguage="Swift" userDefinedModelVersionIdentifier="">
-    <entity name="OfflineFile" representedClassName=".OfflineFile" syncable="YES">
-        <attribute name="downloadDate" optional="YES" attributeType="Date" usesScalarValueType="NO" spotlightIndexingEnabled="YES" syncable="YES"/>
-        <attribute name="localPath" attributeType="String" syncable="YES"/>
-        <attribute name="mime" optional="YES" attributeType="String" syncable="YES"/>
-        <attribute name="mtime" optional="YES" attributeType="Date" usesScalarValueType="NO" syncable="YES"/>
-        <attribute name="name" attributeType="String" spotlightIndexingEnabled="YES" syncable="YES"/>
-        <attribute name="progress" attributeType="Float" defaultValueString="0" usesScalarValueType="YES" syncable="YES"/>
-        <attribute name="remoteFileUri" optional="YES" attributeType="String" syncable="YES"/>
-        <attribute name="size" attributeType="Integer 64" defaultValueString="0" usesScalarValueType="YES" spotlightIndexingEnabled="YES" syncable="YES"/>
-        <attribute name="state" attributeType="Integer 16" defaultValueString="0" usesScalarValueType="YES" syncable="YES"/>
-        <uniquenessConstraints>
-            <uniquenessConstraint>
-                <constraint value="localPath"/>
-            </uniquenessConstraint>
-        </uniquenessConstraints>
-    </entity>
-    <elements>
-        <element name="OfflineFile" positionX="-1206" positionY="-729" width="128" height="180"/>
-    </elements>
-</model>
